Clean up RELEASE_MOVE_VCN_TO_MAINLINE
To move VCN from the platform to Tethering module, the system
defined a RELEASE_MOVE_VCN_TO_MAINLINE flag to indicate whether
this modularization should be enabled. Specifically, this flag
is used in Android.bp files to teach the build system whether
to build VCN into the platform or the module.

This CL cleans up the RELEASE_MOVE_VCN_TO_MAINLINE flag
and all the relevant flag checks in the build files

Bug: 374174952
Test: atest FrameworksVcnTests && atest CtsVcnTestCases
Flag: EXEMPT FLAG_REMOVAL
Change-Id: I543009be2c7c290a21016257b761c9babb6b3605