Include evemu-record command in user builds (attempt 2)
Previously [0] we only added this to userdebug and eng builds, since we
didn't have a compelling reason to add it to user builds. We now do, due
to upcoming dogfood programmes running on user builds.

This is the second attempt at this CL, since the first [1] missed the
required change to product/generic/Android.bp.

[0]: change I21eb7d5dc8cef9c1dd797ba1ab2ee830a8a5a42b
[1]: change I3f73e4b9794da4a28b2b5aa60885e8bde37152a1

Bug: 302297266
Test: on a user build, run the evemu-record command
Test: $ lunch gsi_arm64-next-user
      $ m android_gsi
Flag: UNSUPPORTED_LANGUAGE
Change-Id: I3dab4caeacd460e81c3672a567480e551a852286