add hwservicemanager.always_sets_disabled system prop to base_system
During boot, libhwbinder waits for hwservicemanager to set
hwservicemanager.ready with 1-sec timeout in a loop. However, On devices
with hwservicemanager disabled, .ready is never set, which forces
libhwbinder to wait for the timeout. This slows down the boot time.

To prevent this issue, hwservicemanager will always initialize
hwservicemanager.disabled. That allows libhwbinder to wait for
hwservicemanager.disabled first on all devices.

Since there is no guarantee that libhwbinder and hwservicemanager are
updated together, we add `hwservicemanager.always_sets_disabled` system
property to indicate the existence of this new behavior.

Bug: 419725480
Test: manual (CF); confirmed no timeout messages appear in logcat
Test: manual (CF); boots, manually killing hwservicemanager during boot
Change-Id: I3dbef61855a70500b86cfd4975afb599966a3447