Add UUID to build.prop file in make builds
As part of b/439900398, UUID is being added to build.prop files. The
build.prop have been migrated to soong modules, with the exception of
vendor's build.prop file which is autogenerated by fsgen in soong only
builds.

To provide UUID info when building with Make, add the UUID to the build.prop
file rules in build/make/core

Bug: 445939490
Bug: 439900398
Test: lunch husky-trunk_staging-userdebug
Test: m out/target/product/husky/vendor/build.prop --no-soong-only
Test: verified that uuid is present

Change-Id: I7aa47ff54c4c078a1bffc44faea509128ef0431a