Add mmd_setup binary
Zram setup was previously done by passing an argument to mmd and
directly exec'ing it into the su domain. This was done to avoid giving
unnecessary elevated privileges to the long-running mmd service.
However, that approach doesn't work on user builds, since the su domain
isn't present. To fix this, the mmd binary is being split into two parts
- one for setup and one for the service - each with their own domain.

Bug: 448210809
Test: manually verify mmd works on a user build
Change-Id: I03691eec0e2132304a3afb0cc5d31a498c95806a