Add custom build tool for PPD DevX.
Adding a new tool that can be invoked as part of the SDK generation process to include a custom PPD artifact file as part of the SDK zip to power PPD DevX work in Studio. This the MVP version of the tool that is sufficient for Android 17.

At a high-level, the tool is designed to:
1. Scan the platform manifest XML for permissions that require purpose
   (specific or purposeString variants).
2. Check if the permission is enabled in case it's guarded by the flag.
3. Transform the purpose-guarded permission element along with the valid
   purpose tags into a new XML with SDK versioning info.
4. Write the new XML into a new file.

Refer to other CLs in the topic on how the tool is invoked with
appropriate input parameters.

Bug: 448885868
Test: atest permission_versions_tests; m sdk and verified artifact file.
Flag: HOST_MODE
Change-Id: I79de7ad0d5ca95a1284f1c2962e8aa12c929a65d