Allow desktop firmware module to be configured
This change makes the desktop firmware module configurable by using the DESKTOP_FIRMWARE_MODULE variable.
If the variable is not set, it defaults to desktop-firmware-squashfs to maintain backward compatibility.

This allows device-specific configurations to select the appropriate firmware image (e.g., erofs or squashfs) without modifying the core build files.

This change is accompanied by a change in device/google/desktop/ to set the DESKTOP_FIRMWARE_MODULE variable.

Test: m dist && file out/dist/*-firmware.img
Bug: 454782245
Change-Id: I5ce18c221e87ae0abaf9d826a9d9054641c8f7db