Add @ChecksSdkIntAtLeast for finalized flags
This allows the linter to supress NewApi checks when code is guarded by a finalized flag.

This uses androidx.annotation as the android.annotation version is not available outside of the platform. This is ok since the only consumer of ExportedFlags is outside of the platform. Will use android.annotation for the platform version of the flags in a follow up change.

Changes test_generate_java_code_exported_flags to conditionally use different golden files based on the cfg value. This works with how the golden files are generated too. Questionable how well it scales for many cfg values.

Test: atest aconfig.test
Test: cargo test --features generate_checks_sdk_annotation
Bug: 446846210
Flag: RELEASE_ACONFIG_GENERATE_CHECKS_SDK_ANNOTATION
Change-Id: Ia7d49e39bc5b97cea98479f468d59c8fa8fe6727