Specify profiles for bips preinstall
Instead of using the install-in user-type="PROFILE" catch-all, explicitly
state all of the profiles android.com.bips should be preinstalled on. This
CL removes preinstallation of the print service on supervising profile and
any OEM profiles that may exist.

The print service is necessary for all full users as well as the
managed, private and clone profiles. Each of these users needs the ability to
print.

Bug: 454974025
Test: adb shell dumpsys package com.android.bips | grep "installed=true"
Change-Id: I07045bf3ddaa5c30063d53f2430e47ad70fde86c