Restore pmg to the base system image
The pmg daemon is included in the base system image if the flag is
true.  The daemon was accidentally removed by an earlier commit.

Flashed a local build that set the flag true, and verified that
the pmg daemon (named "pmgd") was present on the phone.

Flag: RELEASE_PROCESS_MEMORY_GUARDIAN_DAEMON
Bug: 434681786
Test: presubmit
Change-Id: I0ffbf108f9baf8e1a428944552d722854c16da49