releasetools: fix undefined salt for empty images
_SignBootableImage() tries to use kernel or ramdisk contents to create a
deterministic salt for AVB signing, but in the current logic if a boot
image didn't contain a kernel or ramdisk the `salt` variable would be
undefined, resulting in an undefined variable exception.

This CL defaults `salt` to None first to allow the function to proceed
in this case. The specific use case is to allow building a
placeholder empty init_boot image with no kernel or ramdisk.

Bug: b/459799623
Test: `m` and `m dist`
Change-Id: Ic833a7f183d3fbf2008cd476e425481483bc1f68