Fix issue in flag value files v4
Checked that when reading a boolean you didn't go into the int section, but when creating the object if there were no ints we just set int offset to 0 (so the boolean offset was always too large). If no ints, set int offset to end of file.

Change-Id: Ia339d38f85f3b357dd78a0605cfc63cd40ef2498
Bug: 439864800
Flag: build.RELEASE_ACONFIG_PARSE_V4