[CredMan][Add Allowlist across Android Users]
CredentialManager is a core, user-facing service that facilitates
credential handling between apps and providers. It is needed for all
interactive users and within specific profiles.

The package is allowlisted for the following user types:
- `FULL`: For all full human users (primary, secondary, guest).
- `android.os.usertype.profile.MANAGED`: For work profiles.
- `android.os.usertype.profile.PRIVATE`: For the Private Space feature.
- `android.os.usertype.profile.CLONE`: For cloned profiles.

This ensures the service is available where needed while preventing it
from being installed on user types where it is not required in the future (via granular removal) - even if currently the idea is that any surface that could login needs it, future changes could render us needing more granular control, hence why we list them out in this format.

Bug: 451959977
Test: Manually tested on a device prior to submission to ensure correctness.
Flag: NONE config change only
Change-Id: Iac9bebc3f0d702fcf29467afda10e150e666e14b