Dist signapk.jar for sdk-repo-build-tools
The prebuilt signapk.jar is stored in prebuilts/sdk/tools/lib,
but it is not distributed as part of the SDK so it has historically
been manually updated.  Dist it as part of sdk-repo-build-tools
so that the prebuilts update script can update it alongside the rest
of the official SDK tools.

Bug: 325130426
Test: manual
Change-Id: I4bd239eab32cb097b59bb09a32ce9b642bb208a8