Update deprecated_at_birth_allowlist.txt
New allowlist is generated from running `sdk-artifact-verifier --api-versions-path $ANDROID_BUILD_TOP/prebuilts/sdk/36.1/public/data/api-versions.xml` and it is compatible with ToT api-versions.xml. This is what we will use in the CI job when it is set up. Moving forward there should be no new entries added to this list (unless metalava changes and we have to work around that).

Test: sdk-artifact-verifier  --api-versions-path out/dist/data/api-versions.xml  --deprecated-at-birth-allowlist-path  $ANDROID_BUILD_TOP/build/make/tools/finalization/sdk-artifact-verifier/deprecated_at_birth_allowlist.txt
Bug: 447587058
Flag: EXEMPT HOST_MODE
Change-Id: I0ad80535d1a5b4a77b6c289194de21889f6621f9