Add max_sdk_version_full flag
Created pub struct MajorMinorVersion and parsing all relevant fields into it. This will make the parsing failed if it encounters any malformed fields.

I've tried to add error output to point out which class/field/method in api-versions.xml that is malformed. Ideally this would print the xml line and content, but that is a quest for another day.

Bug: 447587058
Test: sdk-artifact-verifier --api-versions-path out/dist/data/api-versions.xml --deprecated-at-birth-allowlist-path $ANDROID_BUILD_TOP/build/make/tools/finalization/sdk-artifact-verifier/deprecated_at_birth_allowlist.txt
Flag: EXEMPT HOST_MODE
Change-Id: I33d0fd2cd0663c89ae3d814d9fa3f9c3ad8cc3d6