Revert^3 "Do not generate symbols for make prebuilt module"
This reverts commit b494bb2fde9cb2ce9d2ac7761efdae1b18be1384.

Change from the previous revert is that now whether the symbols path
will be written by a Soong module or not is checked. Prior to
Soong-only, if more than one modules (e.g. one Soong module via
Androidmk entries and one Make module) copies the symbols to the same
path, the last one was used by kati instead of throwing a
duplicte build rule error.

This became a duplicate build rule error when the symbols copy logic was
migrated to Soong, which is why the symbols generation for the make
prebuilt modules feature was removed.

To restore the support, now the symbols path under symbols/ directory is
stored, and the make prebuilt module's symbol is copied only when it is
a unique path under the symbols/ directory. If the path is not unique
and conflicts with the Soong module's symbols path, this can be resolved
by setting a different `stem` property or disabling the conflicting
Soong module.

Reason for revert: Trying resubmission of the revert
Bug: 427867754
Bug: 409181125
Test: lunch <conflicting product>-next-userdebug && m nothing

Change-Id: I80e54de45691e7390a510db5f395c455eb70fa4d