Remove anomaly detector double flagging.
The anomaly detector code is already feature-flagged by the
“android.os.profiling.anomaly.flags.anomaly_detector_core” aconfig flag.

An additional build flag was introduced for the following two reasons:
- Our code is inside the Profiling module and we wanted to avoid shipping it
  in the initial phase;
- We wanted to avoid potential coverage drops in the mainline module due to
  the addition of new code.

These two justifications are no longer valid.
“Profiling” won’t ship a new version until 26Q2 (when we are releasing our
code) and the coverage infrastructure is not smart enough to automatically
exclude code based on a build flag.
We have additionally verified that our code coverage meets the mainline module
target.

This “double flagging” is causing several issues, especially when they go out
of sync.
This feature as a whole should only be behind a single flag.

Bug: 470415713
Test: atest CtsAnomalyDetectorTests
Test: atest MtsAnomalyDetectorTests
Flag: EXEMPT HOST_MODE

Change-Id: I458e3dce628a3295ed045dfbca8d66f841ff3f3a