releasetools: Make .map file optional for non-A/B OTAs
The `GenerateNonAbOtaPackage()` function fails to create a non-A/B OTA
package if a partition's `.map` file is not present in the target_files
zip.

With recent changes to erofs userspace build utilities, `.map` files
may not be generated for erofs partitions.  This causes an assertion
 failure in `GetSparseImage` and `GetNonSparseImage` in `common.py`,
which blocks the generation of non-A/B OTA packages.

This change makes the presence of `.map` files optional:
- In `GetNonSparseImage`, the check for the `.map` file is removed as it
  was not being used.
- In `GetSparseImage`, it now checks for the existence of the `.map` file.
  If the file is not present, it passes `None` to the `SparseImage`
  constructor, which can handle the missing file gracefully.

Bug: 470837901
Test: abtd, th

Change-Id: Ice9286411e9198d6cbb137ab5a5aaacb0f13b6c5
Signed-off-by: Sandeep Dhavale <dhavale@google.com>